added how-to to the main scripts

+++ scripts/build-graph-decode-hyp.sh
@@ -1,36 +1,53 @@
 #!/bin/bash
 
 # Copyright 2017  Speech Lab, EE Dept., IITM (Author: Srinivas Venkattaramanujam)
-set -x
+
+# set -x
+export KALDI_ROOT=/home/theokouz/kaldi/
+source $KALDI_ROOT/tools/config/common_path.sh
 
 . ./longaudio_vars.sh
+
+
+echo "$0 $@"  # Print the command line for logging
+
+if [ $# != 4 ]; then
+  echo "Usage: scripts/build-graph-decode-hyp.sh <num_job> <decode-dir> <working-dir> <log-dir>"
+  echo " e.g.: scripts/build-graph-decode-hyp.sh 16 decode data/working_dir data/working_dir/log_dir"
+  echo "Description: This script creates a decoding graph, performs decoding using decode_fmllr.sh, and matches hypothesis with reference text using SCLITE."
+  exit 1;
+fi
+
+
+
 nj=$1
 decode_dir=$2
 working_dir=$3
 log_dir=$4
-#echo "Executing utils/mkgraph.sh $lang_dir $model_dir $graph_dir" > 
+
+
+
 utils/mkgraph.sh $lang_dir $model_dir $graph_dir >> $log_dir/output.log 2> $log_dir/err.log || exit 1
 rm -rf $model_dir/$decode_dir
 mkdir -p $model_dir/$decode_dir/scoring
-#echo "Executing steps/decode.sh --cmd $decode_cmd  --nj $nj --skip-scoring true $graph_dir $data_dir $model_dir/$decode_dir"
+
 # steps/decode.sh --cmd "run.pl"  --nj $nj --skip-scoring true $graph_dir $data_dir $model_dir/$decode_dir >> $log_dir/output.log 2> $log_dir/err.log || exit 1
 steps/decode_fmllr.sh --cmd "run.pl --mem 2G" --nj 1 --skip_scoring true $graph_dir $data_dir $model_dir/$decode_dir >> $log_dir/output.log 2> $log_dir/err.log || exit 1
-#echo "Making 10.hyp from lattice"
+
 (lattice-scale --inv-acoustic-scale=10 "ark:gunzip -c $model_dir/$decode_dir/lat.*.gz|" ark:- 2> $log_dir/err.log || exit 1)  | \
         (lattice-add-penalty --word-ins-penalty=10.0 ark:- ark:- 2> $log_dir/err.log || exit 1) | \
         (lattice-best-path --word-symbol-table=$lang_dir/words.txt ark:- ark,t:$model_dir/$decode_dir/scoring/10.tra 2> $log_dir/err.log || exit 1 )
 (cat $model_dir/$decode_dir/scoring/10.tra 2> $log_dir/err.log || exit 1) | sed 's/segment_//g' | sort -k1n | cut -d' ' -f2- | tr '\n' ' ' | tr -s ' ' > $working_dir/hypothesis.tra
-#echo "Creating $working_dir/text_ints from $working_dir/text_actual using scripts/sym2int.py"
+
 (python scripts/sym2int.py $lang_dir/words.txt $working_dir/text_actual 2> $log_dir/err.log || exit 1 )| tr -s ' ' > $working_dir/text_ints
-#echo "Creating word_alignment.ctm"
+
 (lattice-add-penalty --word-ins-penalty=10.0 ark:"gunzip -c $model_dir/$decode_dir/lat.*.gz|" ark:- 2> $log_dir/err.log || exit 1)| \
 (lattice-1best  --acoustic-scale=0.1 ark:- ark:- 2> $log_dir/err.log || exit 1) | \
 (lattice-align-words $lang_dir/phones/word_boundary.int $model_dir/final.mdl ark:- ark:- 2> $log_dir/err.log || exit 1) | \
 (nbest-to-ctm ark:- - 2> $log_dir/err.log || exit 1) | sed 's/segment_//g' | sort -s -k 1,1n | sed 's/^/segment_/g' > $working_dir/word_alignment.ctm
+
 # note: multiple spaces before key makes sclite fail and number of lines should match 
-#echo "RM format for  $working_dir/hypothesis.tra"
 (cat $working_dir/hypothesis.tra 2> $log_dir/err.log || exit 1) | sed 's/$/ (key_1)\n/' | tr -s ' '> $working_dir/hypothesis.tra_rm
-#echo "RM format for $working_dir/text_ints"
 (cat $working_dir/text_ints 2> $log_dir/err.log || exit 1) | sed 's/$/ (key_1)\n/' | tr -s ' ' > $working_dir/text_ints_rm
 #echo "Text-Text alignment using sclite"
 sclite -p -i 'rm' -r $working_dir/text_ints_rm -h $working_dir/hypothesis.tra_rm > $working_dir/ref_and_hyp 2> $log_dir/err.log || (echo "sclite failure" && exit 1)
